modularize typdefs and resolvers
