+++ packages/backend/graphql/rootResolvers.js
@@ -1,0 +1,9 @@
+const path = require('path');
+const { mergeResolvers } = require('@graphql-tools/merge');
+const { loadFilesSync } = require('@graphql-tools/load-files');
+
+const resolversArray = loadFilesSync(
+  path.join(__dirname, './**/*.resolvers.js')
+);
+
+module.exports = mergeResolvers(resolversArray);
