+++ packages/backend/graphql/typedefs.js
@@ -1,0 +1,8 @@
+// ./graphql/typeDefs.js
+const path = require('path');
+const { loadFilesSync } = require('@graphql-tools/load-files');
+const { mergeTypeDefs } = require('@graphql-tools/merge');
+
+const typesArray = loadFilesSync(path.join(__dirname, './**/*.schema.js'));
+
+module.exports = mergeTypeDefs(typesArray);
