+++ packages/backend/package.json
@@ -22,6 +22,8 @@
   },
   "homepage": "https://github.com/duncanjbain/bugtracker#readme",
   "dependencies": {
+    "@graphql-tools/load-files": "^6.2.5",
+    "@graphql-tools/merge": "^6.2.7",
     "apollo-server-express": "^2.19.0",
     "bcrypt": "^5.0.0",
     "dotenv": "^8.2.0",
